requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
aiohttp==3.9.1
python-jose==3.3.0
jinja2==3.1.2
python-multipart==0.0.6
redis==5.0.1
cryptography==41.0.7
pydantic==2.5.3
airtable-python-wrapper>=0.15.3
anthropic>=0.19.1
argparse>=1.4.0
python-dotenv>=1.0.0
pydantic>=2.4.2
requests>=2.31.0
typing-extensions>=4.7.1
websockets>=11.0.3
mcp>=1.22.0
fastmcp>=2.13.3